fix: non-thread-safe CreateOrJoinGame 💾
the operation of creating a game or challenge was not atomic, resulting in a lack of thread safety. for example, A doesn't see any challenges, so they create their own challenge. B does the same thing. as a result, after creating their own challenges, A and B will never play each other while waiting, because they called CreateOrJoinGame almost at the same time.