feat: new HistoryManager 👑
in the previous version, we created a new Chess instance, and to move to another ply, we had to reset the board and play all plies up to the desired one, which takes O(ply).

now we don’t create a new Chess instance, but just keep a history array. so to move to a ply, we just set the board to the position from the history array by ply index, which takes O(1).